Changed weights to doubles and moved implementations of functions to cpp.

+++ src/utils.hpp
@@ -14,7 +14,7 @@
 typedef std::vector<std::vector<int>> graf_t;
 
 /// Represents a directed weighted graph.
-typedef std::vector<std::vector<std::pair<int, int>>> weighted_graf_t;
+typedef std::vector<std::vector<std::pair<int, double>>> weighted_graf_t;
 
 /// Represents infinity.
 const double INFINITY = std::numeric_limits<double>::infinity();
